+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v16.0.2...HEAD)
+### Added
+* [PCM-2191](https://inindca.atlassian.net/browse/PCM-2191) - add meetingId property to the pendingSession and the mediaSession objects
+
 # [v16.0.2](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v16.0.1...v16.0.2)
 ### Added
 * [PCM-2091](https://inindca.atlassian.net/browse/PCM-2091) - Be able to handle a duplicate channel message from hawk signaling a mandatory hard reconnect